+++ src/ArchiveFormats/NitroPlus/mpk.h
@@ -13,8 +13,8 @@ class MPKFormat : public ArchiveFormat {
     std::string description = "Nitro+ Resource Archive";
     uint32_t sig = 0x4B504D;
 
-    ArchiveBase* TryOpen(unsigned char *buffer, uint32_t size, std::string file_name) override;
-    bool CanHandleFile(unsigned char *buffer, uint32_t size, const std::string &ext) const override;
+    ArchiveBase* TryOpen(unsigned char *buffer, uint64_t size, std::string file_name) override;
+    bool CanHandleFile(unsigned char *buffer, uint64_t size, const std::string &ext) const override;
     std::string GetTag() const override {
         return this->tag;
     }
