+++ src/GUI/Utils.h
@@ -8,6 +8,6 @@ class Utils {
     public:
         static std::string GetLastModifiedTime(const std::string& path);
         static std::string GetFileSize(const fs::path& path);
-        static std::string GetFileSize(uint32_t size);
+        static std::string GetFileSize(uint64_t size);
         static std::string ToLower(const std::string &str);
 };
